feat: support custom components in menu
