+++ src/menu/use-menu.ts
@@ -21,7 +21,7 @@ export interface MenuListController {
 }
 
 export interface MenuItemController {
-  ref: React.MutableRefObject<HTMLLIElement | null>;
+  ref: React.MutableRefObject<HTMLElement | null>;
   hasSubmenu: boolean;
   onSelect?: (middleButton?: boolean) => void;
 }
